+++ Screens/StartScreen.py
@@ -1,7 +1,7 @@
 import tkinter as tk
 import sys
 import os
-SCRIPT_DIR = os.path.dirname(os.path.abspath("tkinter_objects\__init__.py"))
+SCRIPT_DIR = os.path.dirname(os.path.abspath("tkinter_objects\\__init__.py"))
 sys.path.append(os.path.dirname(SCRIPT_DIR))
 from tkinter_objects.ControllerMainObject import *