+++ Screens/ViewRecipeScreen.py
@@ -2,7 +2,7 @@ import tkinter as tk
 from tkinter import ttk
 import sys
 import os
-SCRIPT_DIR = os.path.dirname(os.path.abspath("tkinter_objects\__init__.py"))
+SCRIPT_DIR = os.path.dirname(os.path.abspath("tkinter_objects\\__init__.py"))
 sys.path.append(os.path.dirname(SCRIPT_DIR))
 
 from tkinter_objects.ControllerMainObject import *
@@ -10,54 +10,58 @@ from TableObjects import *
 
 class ViewRecipe(ChildFrame):
     def __init__(self, master, controller, view_else_create, **kwargs) -> None:
-        # info
+        super().__init__(master, controller, **kwargs)
+        # Variables
         self.recipe:Recipe
         self.category_dictionary = {}
+        self.show_hidden = tk.BooleanVar()
+        self.name_variable = tk.StringVar()
+        self.food_category = tk.StringVar()
         
-        super().__init__(master, controller, **kwargs)
+        # Widgits
         if view_else_create:
             tk.Label(self, text="View Recipe").grid(row=0, column= 0)
         else:
             tk.Label(self, text="Create Recipe").grid(row=0, column= 0)
 
-        self.show_hidden = tk.BooleanVar()
         self.show_hidden_check_button = tk.Checkbutton(self, text="Hide Recipe", variable=self.show_hidden)
         self.show_hidden_check_button.grid(row=1, column=0)
 
         self.name_label = tk.Label(self, text="Name")
         self.name_label.grid(row=2, column=0)
 
-        self.name_variable = tk.StringVar()
         self.name_entry = tk.Entry(self, textvariable=self.name_variable)
         self.name_entry.grid(row = 3, column=0)
 
         self.category_label = tk.Label(self, text="<Category Name>")
         self.category_label.grid(row=4, column=0)
         
-        self.food_category = tk.StringVar()
         self.category_combo_box = ttk.Combobox(self, textvariable=self.food_category)
         self.category_combo_box.grid(row=5, column=0)
         
         #################################################################################################
         ###################################### Ingredient controls ######################################
         #################################################################################################
+        # variables
         self.ingredient = Ingredient()
         self.ingredients = []
+        self.ingredients_to_delete = []
+        self.amount_name_var = tk.StringVar()
+        self.ingredient_name_var = tk.StringVar()
 
+        # Widgits
         self.sub_title_frame = tk.Frame(self)
         self.sub_title_frame.grid(row=6, column=0)
         
         tk.Label(self.sub_title_frame, text="Ingredient List").grid(row=0, column= 0, columnspan=4)
 
-        self.ingredient_name_var = tk.StringVar()
         self.ingredient_name_entry = tk.Entry(self.sub_title_frame, textvariable=self.ingredient_name_var)
         self.ingredient_name_entry.grid(row=1, column=1)
 
-        self.amount_name_var = tk.StringVar()
         self.amount_name_entry = tk.Entry(self.sub_title_frame, textvariable=self.amount_name_var)
         self.amount_name_entry.grid(row=1, column=2)
         
-        self.create_ingredient_button = tk.Button(self.sub_title_frame, text="Add Ingredient")
+        self.create_ingredient_button = tk.Button(self.sub_title_frame, text="Add Ingredient", command=self.save_ingredient)
         self.create_ingredient_button.grid(row=1, column=3)
 
         self.ingredient_list_box = tk.Listbox(self, height=10)
@@ -72,7 +76,7 @@ class ViewRecipe(ChildFrame):
         self.deselect_ingredient_button = tk.Button(self.ingredient_control_frame, text="deselect", command=self.deselect_ingredient)
         self.deselect_ingredient_button.grid(row=0, column=1)
 
-        self.delete_ingredient_button = tk.Button(self.ingredient_control_frame, text="delete ingredient")
+        self.delete_ingredient_button = tk.Button(self.ingredient_control_frame, text="delete ingredient", command=self.delete_ingredient)
         self.delete_ingredient_button.grid(row=0, column=2)
         #################################################################################################
         ###################################### Ingredient controls ######################################
@@ -93,10 +97,10 @@ class ViewRecipe(ChildFrame):
         self.control_frame.grid(row=12)
 
         if view_else_create:
-            self.update_button = tk.Button(self.control_frame, text="Update", command=lambda: controller.open_frame("recipe index"))
+            self.update_button = tk.Button(self.control_frame, text="Update", command= self.save_recipe)
             self.update_button.grid(row= 0, column=0)
         else:
-            self.update_button = tk.Button(self.control_frame, text="Create", command=lambda: controller.open_frame("recipe index"))
+            self.update_button = tk.Button(self.control_frame, text="Create", command= self.save_recipe)
         self.update_button.grid(row= 0, column=0)
 
         self.back_button = tk.Button(self.control_frame, text="Back", command=lambda: controller.open_frame("recipe index"))
@@ -105,7 +109,15 @@ class ViewRecipe(ChildFrame):
         # Footer #
         ##########
     
+
     def get_food_categories(self):
+        '''
+        Grabs all of the categories from the database then creates a list of category names and and creates a dictionary of using the name of the category
+        and the id of the category.
+        {
+        "Name of Category" : 1
+        }
+        '''
         connection : SqliteConnecter = self.controller.get_connector()
         connection.c.execute("""SELECT * FROM food_categories;""")
         list_of_food_categories = connection.c.fetchall()
@@ -114,13 +126,17 @@ class ViewRecipe(ChildFrame):
         dictionary_of_categories = {}
         for category in list_of_food_categories:
             list_of_category_names.append(category[1])
-            dictionary_of_categories[category[1]] = category[0]
+            dictionary_of_categories[category[1]] = FoodCategory().assign_by_array(category)
         
         self.category_combo_box["values"] = list_of_category_names
         self.category_dictionary = dictionary_of_categories
         self.food_category.set("")
 
-    def set_new(self): 
+
+    def set_new(self):
+        '''
+        Resets the screen.
+        '''
         self.recipe = Recipe()
         self.show_hidden.set(False)
         self.name_variable.set("")
@@ -129,6 +145,7 @@ class ViewRecipe(ChildFrame):
         
         self.ingredient = Ingredient()
         self.ingredients = []
+        self.ingredients_to_delete = []
         self.ingredient_name_var.set("")
         self.amount_name_var.set("")
         self.ingredient_list_box.delete(0, 'end')
@@ -137,8 +154,12 @@ class ViewRecipe(ChildFrame):
         self.instruction_text.delete(0.0, 'end')
 
         self.ingredient_list_box.select_clear(0, 'end')
-    
+
+
     def set_recipe(self, recipe:Recipe):
+        '''
+        sets the recipe screen to the selected recipe form the recipe index menu.
+        '''
         self.set_new()
         connection : SqliteConnecter = self.controller.get_connector()
 
@@ -166,19 +187,61 @@ class ViewRecipe(ChildFrame):
             self.ingredient_list_box.insert(ingredient[0], text)
             self.ingredients.append(Ingredient().assign_by_array(ingredient))
     
+    def update_ingredients(self):
+        self.deselect_ingredient()
+        self.ingredient_list_box.delete(0, 'end')
+        print(self.ingredients)
+        for ingredient in self.ingredients:
+                text = ingredient.name + " " + ingredient.amount
+                self.ingredient_list_box.insert('end', text)
+        
     def save_recipe(self):
-        pass
+        connection : SqliteConnecter = self.controller.get_connector()
 
+        self.recipe.name = self.name_variable.get()
+        self.recipe.hidden = self.show_hidden.get()
+        food_category = self.category_dictionary.get(self.food_category.get())
+        self.recipe.description = self.description_text.get('1.0', 'end')
+        self.recipe.instructions = self.instruction_text.get('1.0', 'end')
+        create_full_recipe(connection.c, food_category, self.recipe, self.ingredients)
+
+        for ingredient in self.ingredients_to_delete:
+            connection.c.execute(f"""DELETE FROM ingredients
+                                 WHERE ingredient_id = {ingredient.id}""")
+        connection.conn.commit()
+        self.controller.open_frame("recipe index")
+        print("done")
+
+    
+    ### Ingredient controls
     def save_ingredient(self):
-        pass
+        self.ingredient.name = self.ingredient_name_var.get()
+        self.ingredient.amount = self.amount_name_var.get()
+        print(self.ingredient.id)
+        if self.ingredient.id == -1:
+            self.ingredients.append(self.ingredient)
+        self.update_ingredients()
 
     def edit_ingredient(self):
+        '''
+        Gets the selected ingredient and sets it to be editable
+        '''
         if self.ingredient_list_box.curselection():
-            print(self.ingredients[self.ingredient_list_box.curselection()[0]])
+            selected_ingredient:Ingredient = self.ingredients[self.ingredient_list_box.curselection()[0]]
+            self.ingredient_name_var.set(selected_ingredient.name)
+            self.amount_name_var.set(selected_ingredient.amount)
+            self.ingredient = selected_ingredient
 
     def deselect_ingredient(self):
+        '''
+        deselects the ingredient and deselecting the ingredient in the listbox
+        '''
         self.ingredient = Ingredient()
+        self.ingredient_name_var.set("")
+        self.amount_name_var.set("")
         self.ingredient_list_box.select_clear(0, 'end')
 
     def delete_ingredient(self):
-        pass
+        if self.ingredient_list_box.curselection():
+            self.ingredients_to_delete.append(self.ingredients.pop(self.ingredient_list_box.curselection()[0]))
+            self.update_ingredients()